Allow authentication logic to be overridden by the user, if desired
